+++ docs/CHANGELOG.md
@@ -22,7 +22,7 @@ Entries inside each section should be ordered by type:
 ## CLI
 
 ## Catalog, Lambdas
-* [Added] Add "Qurator" helper for summarizing files content using Bedrock API  ([#3989](https://github.com/quiltdata/quilt/pull/3989))
+* [Added] Add "Qurator" helper for summarizing files content using Bedrock API ([#3989](https://github.com/quiltdata/quilt/pull/3989))
 * [Changed] Use promises for URLs in IGV to have fresh signing each time they used ([#3979](https://github.com/quiltdata/quilt/pull/3979))
 
 # 6.0.0a3 - 2024-04-25